+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientTest.kt
@@ -5,7 +5,6 @@ import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.TokenRequestParameters
-import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.testballoon.invoke
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.eupidsdjwt.EuPidSdJwtScheme
@@ -209,7 +208,7 @@ val OpenId4VciClientTest by testSuite {
                 ),
                 oauth2Client = OAuth2KtorClient(
                     engine = mockEngine,
-                    loadInstanceAttestation = {
+                    loadInstanceAttestation = { _ ->
                         catching {
                             BuildClientAttestationJwt(
                                 SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
@@ -224,7 +223,7 @@ val OpenId4VciClientTest by testSuite {
                             BuildClientAttestationPoPJwt(
                                 SignJwt(clientAuthKeyMaterial, JwsHeaderNone()),
                                 clientId = clientId,
-                                audience = publicContext,
+                                audience = it.authorizationServer,
                                 lifetime = 10.minutes,
                             )
                         }